Move landing page header to landing page
